[Refactor:Site] Split ElectronicGraderController
