+++ site/app/controllers/grading/ElectronicGraderController.php
@@ -13,12 +13,10 @@ use app\models\gradeable\GradedComponent;
 use app\models\gradeable\GradedGradeable;
 use app\models\gradeable\LateDayInfo;
 use app\models\gradeable\LateDays;
-use app\models\gradeable\Mark;
 use app\models\gradeable\Submitter;
 use app\models\gradeable\TaGradedGradeable;
 use app\models\GradeableAutocheck;
 use app\libraries\Logger;
-use app\models\SimpleStat;
 use app\models\GradingOrder;
 use app\models\User;
 use app\libraries\FileUtils;
@@ -52,203 +50,6 @@ class ElectronicGraderController extends AbstractController {
         }
         return true;
     }
-
-    /**
-     * Gnenerates histogram data needed for the TA stats page
-     * @param GradedGradeable[] $overall_scores
-     * @return array of histogram data
-     */
-    public function generateHistogramData($overall_scores) {
-        $include_grade_override  = ($_COOKIE['include_grade_override'] ?? 'omit') === 'include';
-        $include_bad_submissions = ($_COOKIE['include_bad_submissions'] ?? 'omit') === 'include';
-        $include_null_section    = ($_COOKIE['include_null_section'] ?? 'omit') === 'include';
-        $include_withdrawn       = ($_COOKIE['include_withdrawn_students'] ?? 'omit') === 'include';
-
-        //check overrides here to minimize database calls
-        $overridden_user_ids = [];
-        if (!$include_grade_override && count($overall_scores) > 0) {
-            $g_id = $overall_scores[0]->getGradeable()->getId();
-            $rows = $this->core->getQueries()->getRawUsersWithOverriddenGrades($g_id);
-            foreach ($rows as $r) {
-                $overridden_user_ids[$r['user_id']] = true;
-            }
-        }
-
-        $histogram = [
-            "bTA" => [],
-            "tTA" => [],
-            "bAuto" => [],
-            "runtime" => [],
-            "memory" => [],
-            "submitters" => [],
-            "VerConf" => 0,
-            "noSub" => 0,
-            "noActive" => 0,
-            "GradeInq" => 0,
-            "IncompGrading" => 0,
-            "cancelledSub" => 0
-        ];
-
-        // Iterate through all the Scores
-        foreach ($overall_scores as $ov) {
-            $submitter = null;
-            if ($ov->getTaGradedGradeable() !== null) {
-                $submitter = $ov->getTaGradedGradeable()->getGradedGradeable()->getSubmitter();
-            }
-            else {
-                $submitter = $ov->getAutoGradedGradeable()->getGradedGradeable()->getSubmitter();
-            }
-
-            $reg_section = $submitter->getRegistrationSection();
-
-            // should we include null sections
-            if (!$include_null_section && $reg_section === null) {
-                continue;
-            }
-
-            // should we include withdrawn
-            if (!$include_withdrawn && !$ov->getGradeable()->isTeamAssignment() && $submitter->getUser()->getRegistrationType() === "withdrawn") {
-                continue;
-            }
-
-            $submitter_id = $submitter->getId();
-            // should we include grade overrides
-            if (!$include_grade_override && isset($overridden_user_ids[$submitter_id])) {
-                continue;
-            }
-
-            // calculate if too many late days were used
-            $is_bad_submission = false;
-
-            $auto = $ov->getAutoGradedGradeable();
-            if ($auto !== null) {
-                $graded_gradeable = $auto->getGradedGradeable();
-                $user = $submitter->getUser();
-
-                if ($user !== null && $graded_gradeable !== null) {
-                    $ldi = $this->core->getQueries()
-                    ->getLateDayInfoForUserGradeable($user, $graded_gradeable);
-
-                    if ($ldi !== null) {
-                        $is_bad_submission = ($ldi->getStatus() === LateDayInfo::STATUS_BAD);
-                    }
-                }
-            }
-
-            if (!$include_bad_submissions && $is_bad_submission) {
-                continue;
-            }
-
-            //if the submission is valid, add the data to the histogram data
-            if ($ov->getAutoGradedGradeable()->getHighestVersion() !== 0) {
-                if ($ov->getAutoGradedGradeable()->getActiveVersion() !== 0) {
-                    if (!$ov->getOrCreateTaGradedGradeable()->hasVersionConflict()) {
-                    //do not include version conflicts in autograding histogram ^
-                        $histogram["bAuto"] = array_merge($histogram["bAuto"], [$ov->getAutoGradedGradeable()->getTotalPoints()]);
-                        $metrics = $ov->getAutoGradedGradeable()->getMetrics_Sum();
-                        $histogram["runtime"] = array_merge($histogram["runtime"], [$metrics['runtime']]);
-                        $histogram["memory"]  = array_merge($histogram["memory"], [$metrics['memory']]);
-                        $histogram["submitters"] = array_merge($histogram["submitters"], [$ov->getAutoGradedGradeable()->getSubmitterId()]);
-                    }
-                }
-            }
-
-            // Status counters
-            if (!$ov->getAutoGradedGradeable()->hasSubmission()) {
-                $histogram["noSub"] += 1;
-            }
-            elseif ($ov->getAutoGradedGradeable()->getHighestVersion() !== 0 && $ov->getAutoGradedGradeable()->getActiveVersion() === 0) {
-                $histogram["cancelledSub"] += 1;
-            }
-            elseif ($ov->getAutoGradedGradeable()->getActiveVersion() === 0) {
-                $histogram["noActive"] += 1;
-            }
-            elseif ($ov->getTaGradedGradeable()->anyGrades()) {
-                if ($ov->hasActiveGradeInquiry()) {
-                    $histogram["GradeInq"] += 1;
-                }
-                elseif ($ov->getTaGradedGradeable()->hasVersionConflict()) {
-                    $histogram["VerConf"] += 1;
-                }
-                elseif (!$ov->isTaGradingComplete()) {
-                    $histogram["IncompGrading"] += 1;
-                }
-                else {
-                    $histogram["bTA"] = array_merge($histogram["bTA"], [$ov->getTaGradedGradeable()->getTotalScore() + $ov->getAutoGradedGradeable()->getTotalPoints()]);
-                    $histogram["tTA"] = array_merge($histogram["tTA"], [$ov->getTaGradedGradeable()->getTotalScore()]);
-                }
-            }
-        }
-        return $histogram;
-    }
-
-    /**
-     * Helper function for Randomization
-     * @param Array $student_array
-     * @param int $number_to_grade
-     * @return Array $final_grading_info
-     */
-    private function setRandomizedGraders(array $student_array, int $number_to_grade) {
-        $final_grading_info = [];
-        $graded_array = $student_array;
-        /*n_array_peers : An Array of arrays that holds information on to be graded peers
-        [ [A,B,C,D,E,F], [E,F,A,B,C,D], [C,D,E,F,A,B] ]
-        A grades C and E and is graded by C and E.
-        */
-        $n_array_peers = [];
-        shuffle($student_array);
-        array_push($n_array_peers, $student_array);
-        /*final_grading_info : An Array with clear structure of grading rules for peer grading
-        [ [A,[C,E]],[B,[F,D]], ...]
-        A grades C and E, B grades F and D ..and so on!
-        */
-        $max_offset = count($student_array);
-        $offset_array = [];
-        $temp_offset = [];
-        for ($i = 1; $i < $max_offset; ++$i) {
-            array_push($temp_offset, $i);
-        }
-        /* $offset_array contains randomly chosen offsets.
-        $temp_offset helps to ensure no duplicate offsets exist (By removing already chosen offsets)
-        Upon every random choice of an offset from $temp_offset, the value is removed from it.
-        */
-        for ($i = 0; $i < $number_to_grade; ++$i) {
-            $random_offset = array_rand($temp_offset, 1);
-            array_push($offset_array, $temp_offset[$random_offset]);
-            unset($temp_offset[$random_offset]);
-        }
-        foreach ($offset_array as $element) {
-            $temp_array = $student_array;
-            for ($i = 0; $i < $element; $i++) {
-                array_push($temp_array, array_shift($temp_array));
-            }
-            array_push($n_array_peers, $temp_array);
-        }
-        for ($i = 0; $i < count($n_array_peers[0]); ++$i) {
-            $temp = [];
-            for ($j = 1; $j < count($n_array_peers); ++$j) {
-                array_push($temp, $n_array_peers[$j][$i]);
-            }
-            array_push($final_grading_info, [$n_array_peers[0][$i],$temp]);
-        }
-        return $final_grading_info;
-    }
-
-    /**
-     * Helper function for all grade all in randomized peer assignments
-     * @param Array $student_array
-     * @return Array $final_grading_info
-     */
-    private function setAllGradAllGrading($student_array) {
-        $final_grading_info = [];
-        for ($grader = 0; $grader < count($student_array); ++$grader) {
-            $peer_array = $student_array;
-            unset($peer_array[$grader]);
-            $peer_array = array_values($peer_array);
-            array_push($final_grading_info, [$student_array[$grader],$peer_array]);
-        }
-        return $final_grading_info;
-    }
     /**
      * Route for Getting Gradeable
      *
@@ -259,124 +60,6 @@ class ElectronicGraderController extends AbstractController {
     public function getUserGroup($gradeable_id) {
         $user_group = $this->core->getUser()->getGroup();
         return JsonResponse::getSuccessResponse($user_group);
-    }
-    /**
-     * Route for randomizing peer assignments with 'One Grades Many'
-     */
-    #[AccessControl(role: "INSTRUCTOR")]
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/RandomizePeers", methods: ["POST"])]
-    public function RandomizePeers($gradeable_id) {
-        /* How does this function work?
-        1 - Number of students to grade (Y) is taken from the client using POST
-            1.1 - If the number is > number of students, then ALL grade ALL.
-        2 - Query DB to get students from registration section(X) (Without taking in students in NULL section)
-        3 - Randomize the order of students
-        4 - Randomly Select Y offsets
-        5 - Shift the random order by the offsets to create the matrix, with no duplicates, and exactly Y assignments and & graders for each student.  no student grades self.
-        */
-        $restrict_to_registration = ($_POST['restrict_to_registration'] ?? '') === "checked";
-        $submit_before_grading = ($_POST['submit_before_grading'] ?? '') === "checked";
-        $number_to_grade = $_POST['number_to_grade'] ?? '';
-        if (!is_numeric($number_to_grade)) {
-            $this->core->addErrorMessage("Peer assignment failed: An invalid number of students to grade was assigned");
-            return JsonResponse::getFailResponse("Number of students to grade expects an integer");
-        }
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            $this->core->addErrorMessage('Invalid Gradeable!');
-            return JsonResponse::getFailResponse("Invalid gradeable ID");
-        }
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->addErrorMessage('This gradeable is not an electronic file gradeable');
-            return JsonResponse::getFailResponse('This gradeable is not an electronic file gradeable');
-        }
-        /* If Restrict to Registration checkbox is checked, then the randomised peer assignments should be restricted to each registration section" */
-        if ($restrict_to_registration) {
-            $sections = $this->core->getQueries()->getRegistrationSections();
-            foreach ($sections as $i => $section) {
-                $sections[$i] = $section['sections_registration_id'];
-            }
-            /* Duplicating the entire functionality of randomization, but for each registration section rather than whole class" */
-            $final_grading_info = [];
-            foreach ($sections as $section) {
-                $student_list = [];
-                $all_grade_all_registration = false;
-                $student_array = [];
-                $students = $this->core->getQueries()->getUsersByRegistrationSections([$section]);
-                foreach ($students as $student) {
-                    array_push($student_list, ['user_id' => $student->getId()]);
-                    if ($submit_before_grading) {
-                        if ($this->core->getQueries()->getUserHasSubmission($gradeable, $student->getId())) {
-                            array_push($student_array, $student->getId());
-                        }
-                    }
-                    else {
-                        array_push($student_array, $student->getId());
-                    }
-                }
-
-                $number_of_students = count($student_array);
-                /* If number of students entered is more than number of students in registration section,
-                then for each registration section with less number of students, everyone will grade everyone */
-                if ($number_to_grade >= $number_of_students) {
-                    $all_grade_all_registration = true;
-                }
-                if ($all_grade_all_registration) {
-                    for ($grader = 0; $grader < count($student_array); ++$grader) {
-                        $peer_array = $student_array;
-                        unset($peer_array[$grader]);
-                        $peer_array = array_values($peer_array);
-                        array_push($final_grading_info, [$student_array[$grader],$peer_array]);
-                    }
-                }
-                else {
-                    $final_grading_info = $this->setRandomizedGraders($student_array, $number_to_grade);
-                }
-            }
-            $gradeable->setRandomPeerGradersList($final_grading_info);
-            return JsonResponse::getSuccessResponse($final_grading_info);
-        }
-
-        $all_grade_all = false;
-        $order = new GradingOrder($this->core, $gradeable, $this->core->getUser(), true);
-        $student_array = [];
-        $student_list = [];
-        $students = $this->core->getQueries()->getUsersByRegistrationSections($order->getSectionNames());
-        foreach ($students as $student) {
-            $reg_sec = ($student->getRegistrationSection() === null) ? 'NULL' : $student->getRegistrationSection();
-            $sorted_students[$reg_sec][] = $student;
-            array_push($student_list, ['user_id' => $student->getId()]);
-            if ($submit_before_grading) {
-                if ($this->core->getQueries()->getUserHasSubmission($gradeable, $student->getId())) {
-                    array_push($student_array, $student->getId());
-                }
-            }
-            else {
-                array_push($student_array, $student->getId());
-            }
-        }
-        $number_of_students = count($student_array);
-        if ($number_of_students <= 1) {
-            $this->core->addErrorMessage("Peer assignments failed: Not enough submissions");
-            return JsonResponse::getFailResponse("Not Enough Submissions");
-        }
-        if ($number_to_grade > $number_of_students) {
-            $all_grade_all = true;
-        }
-        if ($all_grade_all) {
-            $final_grading_info = $this->setAllGradAllGrading($student_array);
-            $gradeable->setRandomPeerGradersList($final_grading_info);
-            return JsonResponse::getSuccessResponse($final_grading_info);
-        }
-        $final_grading_info = $this->setRandomizedGraders($student_array, $number_to_grade);
-        if ($number_to_grade < 1) {
-            $gradeable->setRandomPeerGradersList($final_grading_info);
-            return JsonResponse::getSuccessResponse("Clear Peer Matrix");
-        }
-        $gradeable->setRandomPeerGradersList($final_grading_info);
-        $gradeable->setPeerGradingSet($number_to_grade);
-        return JsonResponse::getSuccessResponse($final_grading_info);
     }
     /**
      * Route for getting whitespace information for the diff viewer
@@ -459,536 +142,6 @@ class ElectronicGraderController extends AbstractController {
         $white_spaces = $diff_viewer->getWhiteSpaces();
         return ['html' => $html, 'whitespaces' => $white_spaces];
     }
-
-    /**
-     * Route for verifying the grader of a graded component
-     * @param string $gradeable_id verify all components or not
-     * @param bool $verify_all false be default
-     */
-    #[AccessControl(permission: "grading.electronic.verify_grader")]
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/verify", methods: ["POST"])]
-    public function ajaxVerifyComponent($gradeable_id, $verify_all = false) {
-        $anon_id = $_POST['anon_id'] ?? '';
-
-        $grader = $this->core->getUser();
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-        // Get user id from the anon id
-        $submitter_id = $this->tryGetSubmitterIdFromAnonId($anon_id, $gradeable_id);
-        if ($submitter_id === false) {
-            return;
-        }
-        // Get the graded gradeable
-        $graded_gradeable = $this->tryGetGradedGradeable($gradeable, $submitter_id);
-        if ($graded_gradeable === false) {
-            return;
-        }
-        // Get / create the TA grade
-        $ta_graded_gradeable = $graded_gradeable->getOrCreateTaGradedGradeable();
-        if (!$verify_all) {
-            $component_id = $_POST['component_id'] ?? '';
-            // get the component
-            $component = $this->tryGetComponent($gradeable, $component_id);
-            if ($component === false) {
-                return;
-            }
-
-            // Get / create the graded component
-            $graded_component = $ta_graded_gradeable->getOrCreateGradedComponent($component, $grader, false);
-
-            // Verifying individual component should fail if its ungraded
-            if ($graded_component === null) {
-                $this->core->getOutput()->renderJsonFail('Cannot verify ungraded component');
-                return;
-            }
-        }
-        try {
-            if ($verify_all === 'true') {
-                foreach ($gradeable->getComponents() as $comp) {
-                    if (!$comp->isPeerComponent()) {
-                        $graded_component = $ta_graded_gradeable->getGradedComponent($comp);
-                        if ($graded_component !== null && $graded_component->getGraderId() != $grader->getId()) {
-                            $graded_component->setVerifier($grader);
-                            $graded_component->setVerifyTime($this->core->getDateTimeNow());
-                        }
-                    }
-                }
-            }
-            else {
-                if (!isset($graded_component)) {
-                    throw new \RuntimeException('Graded component should not be null if $verify_all === false');
-                }
-                $graded_component->setVerifier($grader);
-                $graded_component->setVerifyTime($this->core->getDateTimeNow());
-            }
-            $this->core->getQueries()->saveTaGradedGradeable($ta_graded_gradeable);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    /**
-     * Shows statistics for the grading status of a given electronic submission. This is shown to all full access
-     * graders. Limited access graders will only see statistics for the sections they are assigned to.
-     */
-    #[AccessControl(role: "LIMITED_ACCESS_GRADER")]
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/grading/status")]
-    public function showStatus($gradeable_id) {
-        $gradeable = $this->tryGetGradeable($gradeable_id, false);
-        if ($gradeable === false) {
-            $this->core->addErrorMessage('Invalid gradeable id');
-            $this->core->redirect($this->core->buildCourseUrl());
-        }
-
-        if (!$gradeable->hasAutogradingConfig()) {
-            $this->core->getOutput()->renderOutput(
-                'Error',
-                'unbuiltGradeable',
-                $gradeable,
-                "grades"
-            );
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->addErrorMessage('This gradeable is not an electronic file gradeable');
-            $this->core->redirect($this->core->buildCourseUrl());
-        }
-
-        if (!$this->core->getAccess()->canI("grading.electronic.status", ["gradeable" => $gradeable])) {
-            $this->core->addErrorMessage("You do not have permission to grade {$gradeable->getTitle()}");
-            $this->core->redirect($this->core->buildCourseUrl());
-        }
-
-        $gradeableUrl = $this->core->buildCourseUrl(['gradeable', $gradeable->getId(), 'grading', 'details']);
-        $this->core->getOutput()->addBreadcrumb("{$gradeable->getTitle()} Grading", $gradeableUrl);
-        $this->core->getOutput()->addBreadcrumb("Statistics & Charts");
-
-        $isPeerGradeable = false;
-        if ($gradeable->hasPeerComponent() && ($this->core->getUser()->getGroup() < User::GROUP_STUDENT)) {
-            $isPeerGradeable = true;
-        }
-        $peer = false;
-        if ($gradeable->hasPeerComponent() && ($this->core->getUser()->getGroup() == User::GROUP_STUDENT)) {
-            $peer = true;
-        }
-
-        //get all graded gradeables for queue stats
-        $submissions_in_queue = 0;
-        $gradeables[] = $gradeable;
-        $graded_gradeables = $this->core->getQueries()->getGradedGradeables($gradeables);
-        //check every submission (inactive and active) for every student
-        //NOTE: Let's eventually refactor this nested loop to instead look at the items
-        //in the autograding queue and check for matches.
-        foreach ($graded_gradeables as $g) {
-            $highest_version = $g->getAutoGradedGradeable()->getHighestVersion();
-            if ($highest_version > 0) {
-                for ($i = 1; $i < $highest_version + 1; $i++) {
-                    $display_version_instance = $g->getAutoGradedGradeable()->getAutoGradedVersionInstance($i);
-                    if ($display_version_instance->isQueued()) {
-                        $submissions_in_queue += 1;
-                    }
-                }
-            }
-        }
-
-        /*
-         * we need number of students per section
-         */
-
-        $team_users = [];
-        $no_team_users = [];
-        $my_grading = 0;
-        $num_components = 0;
-        $graded_components = [];
-        $late_components = [];
-        $ta_graded_components = [];
-        $graders = [];
-        $average_scores = [];
-        $sections = [];
-        $total_users = [];
-        $component_averages = [];
-        $histogram_data = [];
-        $manual_average = null;
-        $autograded_average = null;
-        $overall_average = null;
-        $overall_scores = null;
-        $order = null;
-        $num_submitted = [];
-        $num_unsubmitted = 0;
-        $total_indvidual_students = 0;
-        $viewed_grade = 0;
-        $num_gradeables = 1;
-        $total_who_submitted = 0;
-        $peers_to_grade = 0;
-        $peer_graded_components = 0;
-        $peer_components = 0;
-        $total_users_who_submitted = [];
-        $graders_of_inquiries = [];
-        $verified_components = [];
-
-        $include_withdrawn_students = ($_COOKIE['include_withdrawn_students'] ?? 'omit') === 'include';
-
-        // Assure late day cache is calculated
-        $this->core->getQueries()->generateLateDayCacheForUsers();
-        $section_key = ($gradeable->isGradeByRegistration() ? 'registration_section' : 'rotating_section');
-        $grade_inquiries = $this->core->getQueries()->getNumberGradeInquiries($gradeable_id, $gradeable->isGradeInquiryPerComponentAllowed());
-        $graders_of_inquiries = $this->core->getQueries()->getGraderofGradeInquiry($gradeable_id, $gradeable->isGradeInquiryPerComponentAllowed());
-        if ($isPeerGradeable) {
-            $total_users_who_submitted = $this->core->getQueries()->getTotalSubmittedUserCountByGradingSections($gradeable_id, $sections, $section_key, $include_withdrawn_students);
-            $peer_graded_components = 0;
-            $order = new GradingOrder($this->core, $gradeable, $this->core->getUser(), true);
-            $student_array = [];
-            $student_list = [];
-            $students = $this->core->getQueries()->getUsersByRegistrationSections($order->getSectionNames());
-            foreach ($students as $student) {
-                $reg_sec = ($student->getRegistrationSection() === null) ? 'NULL' : $student->getRegistrationSection();
-                $sorted_students[$reg_sec][] = $student;
-                array_push($student_list, ['user_id' => $student->getId()]);
-                array_push($student_array, $student->getId());
-            }
-            foreach ($student_array as $student) {
-                $peer_graded_components += $this->core->getQueries()->getNumGradedPeerComponents($gradeable_id, $student);
-                $peer_components += count($this->core->getQueries()->getPeerAssignment($gradeable_id, $student));
-            }
-        }
-        if ($peer) {
-            $total_users = $this->core->getQueries()->getTotalUserCountByGradingSections($sections, $section_key, $include_withdrawn_students);
-            $peer_array = $this->core->getQueries()->getPeerAssignment($gradeable_id, $this->core->getUser()->getId());
-            $peers_to_grade = count($peer_array);
-            $num_components = count($gradeable->getPeerComponents());
-            $graded_components = $this->core->getQueries()->getGradedPeerComponentsByRegistrationSection($gradeable_id, $sections);
-            $late_components = $this->core->getQueries()->getBadTeamSubmissionsByGradingSection($gradeable_id, $sections, $section_key);
-            $ta_graded_components = $this->core->getQueries()->getGradedPeerComponentsByRegistrationSection($gradeable_id, $sections);
-            $num_gradeables = count($this->core->getQueries()->getPeerGradingAssignmentsForGrader($this->core->getUser()->getId()));
-            $my_grading = $this->core->getQueries()->getNumGradedPeerComponents($gradeable_id, $this->core->getUser()->getId());
-            $component_averages = [];
-            $manual_average = null;
-            $autograded_average = null;
-            $overall_average = null;
-            $overall_scores = null;
-        }
-        elseif ($gradeable->isGradeByRegistration()) {
-            if (!$this->core->getAccess()->canI("grading.electronic.status.full")) {
-                $sections = $this->core->getUser()->getGradingRegistrationSections();
-            }
-            else {
-                $sections = $this->core->getQueries()->getRegistrationSections();
-                foreach ($sections as $i => $section) {
-                    $sections[$i] = $section['sections_registration_id'];
-                }
-            }
-            if (count($sections) > 0) {
-                $graders = $this->core->getQueries()->getGradersForRegistrationSections($sections);
-            }
-            $num_components = count($gradeable->getNonPeerComponents());
-        }
-        else {
-            //grading by rotating section
-            if (!$this->core->getAccess()->canI("grading.electronic.status.full")) {
-                $sections = $this->core->getQueries()->getRotatingSectionsForGradeableAndUser($gradeable_id, $this->core->getUser()->getId());
-            }
-            else {
-                $sections = $this->core->getQueries()->getRotatingSections();
-                foreach ($sections as $i => $section) {
-                    $sections[$i] = $section['sections_rotating_id'];
-                }
-            }
-            if (count($sections) > 0) {
-                $graders = $this->core->getQueries()->getGradersForRotatingSections($gradeable_id, $sections);
-            }
-        }
-
-        //Check if this is a team project or a single-user project
-        if ($gradeable->isTeamAssignment()) {
-            $num_submitted = $this->core->getQueries()->getSubmittedTeamCountByGradingSections($gradeable_id, $sections, $section_key);
-            $late_submitted = $this->core->getQueries()->getBadTeamSubmissionsByGradingSection($gradeable_id, $sections, $section_key);
-        }
-        else {
-            $num_submitted = $this->core->getQueries()->getTotalSubmittedUserCountByGradingSections($gradeable_id, $sections, $section_key, $include_withdrawn_students);
-            $late_submitted = $this->core->getQueries()->getBadUserSubmissionsByGradingSection($gradeable_id, $sections, $section_key);
-        }
-
-        if (count($sections) > 0) {
-            if ($gradeable->isTeamAssignment()) {
-                $total_users = $this->core->getQueries()->getTotalTeamCountByGradingSections($gradeable_id, $sections, $section_key);
-                $no_team_users = $this->core->getQueries()->getUsersWithoutTeamByGradingSections($gradeable_id, $sections, $section_key);
-                $team_users = $this->core->getQueries()->getUsersWithTeamByGradingSections($gradeable_id, $sections, $section_key);
-                $individual_viewed_grade = $this->core->getQueries()->getNumUsersWhoViewedTeamAssignmentBySection($gradeable, $sections);
-            }
-            else {
-                $total_users = $this->core->getQueries()->getTotalUserCountByGradingSections($sections, $section_key, $include_withdrawn_students);
-                $no_team_users = [];
-                $team_users = [];
-                $individual_viewed_grade = 0;
-            }
-            $override_cookie = $_COOKIE['include_grade_override'] ?? 'omit';
-            $bad_submissions_cookie = $_COOKIE['include_bad_submissions'] ?? 'omit';
-            $null_section_cookie = $_COOKIE['include_null_section'] ?? 'omit';
-            $graded_components = $this->core->getQueries()->getGradedComponentsCountByGradingSections($gradeable_id, $sections, $section_key, $gradeable->isTeamAssignment(), $include_withdrawn_students);
-            $late_components = $this->core->getQueries()->getBadGradedComponentsCountByGradingSections($gradeable_id, $sections, $section_key, $gradeable->isTeamAssignment(), $include_withdrawn_students);
-            $component_averages = $this->core->getQueries()->getAverageComponentScores($gradeable_id, $section_key, $gradeable->isTeamAssignment(), $bad_submissions_cookie, $null_section_cookie, $include_withdrawn_students);
-            $autograded_average = $this->core->getQueries()->getAverageAutogradedScores($gradeable_id, $section_key, $gradeable->isTeamAssignment(), $bad_submissions_cookie, $null_section_cookie, $include_withdrawn_students);
-            $overall_average = $this->core->getQueries()->getAverageForGradeable($gradeable_id, $section_key, $gradeable->isTeamAssignment(), $override_cookie, $bad_submissions_cookie, $null_section_cookie, $include_withdrawn_students);
-            $manual_average = new SimpleStat($this->core, [
-                'avg_score' => $overall_average['manual_avg_score'],
-                'std_dev' => $overall_average['manual_std_dev'],
-                'max' => $overall_average['manual_max_score'],
-                'count' => $overall_average['count']
-            ]);
-            $overall_average = new SimpleStat($this->core, $overall_average);
-            $order = new GradingOrder($this->core, $gradeable, $this->core->getUser(), true);
-            $overall_scores = [];
-            $overall_scores = $order->getSortedGradedGradeables();
-            $num_components = count($gradeable->getNonPeerComponents());
-            $viewed_grade = $this->core->getQueries()->getNumUsersWhoViewedGradeBySections($gradeable, $sections, $null_section_cookie);
-            $histogram_data = $this->generateHistogramData($overall_scores);
-            $verified_components = $this->core->getQueries()->getVerifiedComponentsCountByGradingSections($gradeable_id, $sections, $section_key, $gradeable->isTeamAssignment());
-        }
-        $sections = [];
-        //Either # of teams or # of students (for non-team assignments). Either case
-        // this is the max # of submitted copies for this gradeable.
-        $total_submissions = 0;
-        if (count($total_users) > 0) {
-            foreach ($total_users as $key => $value) {
-                // If we allow NULL sections, use any.
-                // If not, make sure $key is not NULL
-                if ($key === "NULL" && ($_COOKIE['include_null_section'] ?? 'omit') === 'omit') {
-                    continue;
-                }
-                $total_submissions += $value;
-            }
-            foreach ($total_users_who_submitted as $key => $value) {
-                // If we allow NULL sections, use any.
-                // If not, make sure $key is not NULL
-                if ($key === "NULL" && ($_COOKIE['include_null_section'] ?? 'omit') === 'omit') {
-                    continue;
-                }
-                $total_who_submitted += $value;
-            }
-            if (!$gradeable->isTeamAssignment() && $isPeerGradeable) {
-                $sections['peer_stu_grad'] = [
-                   // Total peer components to grade = Number of peer components * Number of Students who submitted
-                 'total_who_submitted' => $total_who_submitted,
-                 'total_components' => count($gradeable->getPeerComponents()) * $total_who_submitted,
-                 'non_late_total_components' => count($gradeable->getPeerComponents()) * $total_who_submitted,
-                 'graded_components' => 0,
-                 'verified_components' => 0,
-                 'non_late_graded_components' => 0,
-                 'non_late_verified_components' => 0,
-                 'view_peer_graded_components' => $peer_graded_components,
-                 'view_peer_components' => $peer_components,
-                 'ta_graded_components' => 0,
-                 'num_gradeables' => $num_gradeables,
-                 'graders' => [],
-                 'valid_graders' => []
-                ];
-            }
-            if ($peer) {
-                 // If a team assignment => Team Peer Grading Stats Should be Visible
-                 // Stats are broken, Update this after Teams work fine with Randomized Peer Assignments
-                if ($gradeable->isTeamAssignment()) {
-                    $sections['stu_grad'] = [
-                        'total_components' => count($gradeable->getPeerComponents()), // Multiply it by number of teams assigned to grade
-                        'non_late_total_components' => count($gradeable->getPeerComponents()),
-                        'graded_components' => $my_grading,
-                        'verified_components' => $my_grading,
-                        'non_late_graded_components' => $my_grading,
-                        'non_late_verified_components' => $my_grading,
-                        'num_gradeables' => $num_gradeables,
-                        'ta_graded_components' => 0,
-                        'graders' => [],
-                        'valid_graders' => []
-                    ];
-                    $sections['all'] = [
-                        'total_components' => 0,
-                        'graded_components' => 0,
-                        'verified_components' => 0,
-                        'non_late_total_components' => 0,
-                        'non_late_graded_components' => 0,
-                        'non_late_verified_components' => 0,
-                        'graders' => [],
-                        'valid_graders' => []
-                    ];
-                    foreach ($total_users as $key => $value) {
-                        if ($key == 'NULL') {
-                            continue;
-                        }
-                        $sections['all']['total_components'] += $value * $num_components;
-                        $sections['all']['graded_components'] += isset($graded_components[$key]) ? $graded_components[$key] : 0;
-                        $sections['all']['verified_components'] += $verified_components[$key] ?? 0;
-                    }
-                    $sections['all']['total_components'] -= $num_components;
-                    $sections['all']['graded_components'] -= $my_grading;
-                    $sections['all']['verified_components'] -= $my_grading;
-                    $sections['all']['non_late_total_components'] = $sections['all']['total_components'];
-                    $sections['all']['non_late_graded_components'] = $sections['all']['graded_components'];
-                    $sections['all']['non_late_verified_components'] = $sections['all']['verified_components'];
-                    $sections['stu_grad']['no_team'] = 0;
-                    $sections['stu_grad']['team'] = 0;
-                    $sections['all']['no_team'] = 0;
-                    $sections['all']['team'] = 0;
-                }
-                else {
-                    // If not a team assignment => Individual Peer Grading Stats Should be Visible
-                    $sections['stu_grad'] = [
-                        'total_components' => $num_components * $peers_to_grade,
-                        'non_late_total_components' => $num_components * $peers_to_grade,
-                        'graded_components' => $my_grading,
-                        'verified_components' => $my_grading,
-                        'non_late_graded_components' => $my_grading,
-                        'non_late_verified_components' => $my_grading,
-                        'num_gradeables' => $num_gradeables,
-                        'ta_graded_components' => 0,
-                        'graders' => [],
-                        'valid_graders' => []
-                    ];
-                    $sections['all'] = [
-                        'total_components' => 0,
-                        'graded_components' => 0,
-                        'verified_components' => 0,
-                        'non_late_total_components' => 0,
-                        'non_late_graded_components' => 0,
-                        'non_late_verified_components' => 0,
-                        'graders' => [],
-                        'valid_graders' => []
-                    ];
-                    foreach ($total_users as $key => $value) {
-                        if ($key == 'NULL') {
-                            continue;
-                        }
-                        $sections['all']['total_components'] += $value * $num_components;
-                        $sections['all']['graded_components'] += isset($graded_components[$key]) ? $graded_components[$key] : 0;
-                        $sections['all']['verified_components'] += $verified_components[$key] ?? 0;
-                    }
-                    $sections['all']['total_components'] -= $num_components;
-                    $sections['all']['graded_components'] -= $my_grading;
-                    $sections['all']['verified_components'] -= $my_grading;
-                    $sections['all']['non_late_total_components'] = $sections['all']['total_components'];
-                    $sections['all']['non_late_graded_components'] = $sections['all']['graded_components'];
-                    $sections['all']['non_late_verified_components'] = $sections['all']['verified_components'];
-                }
-            }
-            else {
-                foreach ($total_users as $key => $value) {
-                    if (array_key_exists($key, $num_submitted)) {
-                        $sections[$key] = [
-                            'total_components' => $num_submitted[$key],
-                            'non_late_total_components' => $num_submitted[$key] - (array_key_exists($key, $late_submitted) ? $late_submitted[$key] : 0),
-                            'graded_components' => 0,
-                            'verified_components' => 0,
-                            'non_late_graded_components' => 0,
-                            'non_late_verified_components' => 0,
-                            'ta_graded_components' => 0,
-                            'graders' => [],
-                            'valid_graders' => []
-                        ];
-                    }
-                    else {
-                        $sections[$key] = [
-                            'total_components' => 0,
-                            'graded_components' => 0,
-                            'verified_components' => 0,
-                            'non_late_total_components' => 0,
-                            'non_late_graded_components' => 0,
-                            'non_late_verified_components' => 0,
-                            'graders' => [],
-                            'valid_graders' => []
-                        ];
-                    }
-                    if ($gradeable->isTeamAssignment()) {
-                        $sections[$key]['no_team'] = $no_team_users[$key];
-                        $sections[$key]['team'] = $team_users[$key];
-                    }
-                    if (isset($graded_components[$key])) {
-                        // Clamp to total components if unsubmitted assignment is graded for whatever reason
-                        $sections[$key]['graded_components'] = $graded_components[$key];
-                        $sections[$key]['non_late_graded_components'] = $graded_components[$key] - $late_components[$key];
-                        $sections[$key]['ta_graded_components'] = min(intval($graded_components[$key]), $sections[$key]['total_components']);
-                    }
-                    if (isset($verified_components[$key])) {
-                        $sections[$key]['verified_components'] = $verified_components[$key];
-                        $sections[$key]['non_late_verified_components'] = $verified_components[$key];
-                    }
-                    if (isset($graders[$key])) {
-                        $sections[$key]['graders'] = $graders[$key];
-
-                        // If we allow NULL sections, use any.
-                        // If not, make sure $key is not NULL
-                        if ((array_key_exists('include_null_registration', $_COOKIE) && $_COOKIE['include_null_registration'] === 'true') || $key !== "NULL") {
-                            $valid_graders = [];
-                            foreach ($graders[$key] as $valid_grader) {
-                                /* @var User $valid_grader */
-                                if ($this->core->getAccess()->canUser($valid_grader, "grading.electronic.grade", ["gradeable" => $gradeable])) {
-                                    $valid_graders[] = $valid_grader->getDisplayedGivenName();
-                                }
-                            }
-                            $sections[$key]["valid_graders"] = $valid_graders;
-                        }
-                    }
-                }
-            }
-        }
-        $registered_but_not_rotating = count($this->core->getQueries()->getRegisteredUsersWithNoRotatingSection());
-        $rotating_but_not_registered = count($this->core->getQueries()->getUnregisteredStudentsWithRotatingSection());
-
-        $show_warnings = $this->core->getAccess()->canI("grading.electronic.status.warnings");
-
-        if (isset($order) && $gradeable->isTeamAssignment()) {
-            $total_students_submitted = 0;
-            foreach ($order->getSortedGradedGradeables() as $g) {
-                $team = $g->getSubmitter()->getTeam();
-                $team_section = $gradeable->isGradeByRegistration() ? $team->getRegistrationSection() : $team->getRotatingSection();
-                if (array_key_exists($team_section, $total_users)) {
-                    if ($this->core->getQueries()->getActiveVersionForTeam($gradeable->getId(), $team->getId()) != 0) {
-                        $total_students_submitted += count($team->getMembers());
-                    }
-                }
-            }
-        }
-        else {
-            $total_students_submitted = 0;
-        }
-        $this->core->getOutput()->renderOutput(
-            ['grading', 'ElectronicGrader'],
-            'statusPage',
-            $gradeable,
-            $sections,
-            $component_averages,
-            $manual_average,
-            $autograded_average,
-            $overall_scores,
-            $overall_average,
-            $histogram_data,
-            $total_submissions,
-            $individual_viewed_grade ?? 0,
-            $total_students_submitted,
-            $registered_but_not_rotating,
-            $rotating_but_not_registered,
-            $viewed_grade,
-            $section_key,
-            $grade_inquiries,
-            $graders_of_inquiries,
-            $show_warnings,
-            $submissions_in_queue
-        );
-    }
-
     /**
      * Shows the list of submitters
      */
@@ -2064,102 +1217,6 @@ class ElectronicGraderController extends AbstractController {
         $this->core->getOutput()->renderOutput(['grading', 'ElectronicGrader'], 'popupMarkConflicts');
         $this->core->getOutput()->renderOutput(['grading', 'ElectronicGrader'], 'popupSettings');
     }
-
-    /**
-     * Route for fetching a gradeable's rubric information
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/rubric", methods: ["GET"])]
-    public function ajaxGetGradeableRubric($gradeable_id) {
-        $grader = $this->core->getUser();
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure that this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.grade", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to get gradeable rubric data');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $results = $this->getGradeableRubric($gradeable, $grader);
-            $this->core->getOutput()->renderJsonSuccess($results);
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function getGradeableRubric(Gradeable $gradeable, User $grader) {
-        $return = [
-            'id' => $gradeable->getId(),
-            'precision' => $gradeable->getPrecision()
-        ];
-        // Filter out the components that we shouldn't see
-        //  TODO: instructors see all components, some may not be visible in non-super-edit-mode
-        $return['components'] = array_map(function (Component $component) {
-            return $component->toArray();
-        }, array_filter($gradeable->getComponents(), function (Component $component) use ($gradeable) {
-            return $this->core->getAccess()->canI('grading.electronic.view_component', ['gradeable' => $gradeable, 'component' => $component]);
-        }));
-        $return['components'] = array_values($return['components']);
-        return $return;
-    }
-
-    /**
-     * Gets a component and all of its marks
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components")]
-    public function ajaxGetComponent($gradeable_id, $component_id) {
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-        // Make sure that this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        $graded_gradeable = $this->tryGetGradedGradeable($gradeable, $this->core->getUser()->getId(), false);
-
-        // Get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.view_component", ["gradeable" => $gradeable, "component" => $component])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to get component');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $this->core->getOutput()->renderJsonSuccess($component->toArray());
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
     /**
      * Route for getting information about a individual grader
      */
@@ -2686,518 +1743,6 @@ class ElectronicGraderController extends AbstractController {
             return JsonResponse::getErrorResponse($e->getMessage());
         }
     }
-
-    /**
-     * Route for saving a component's properties (not its marks)
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/save", methods: ["POST"])]
-    public function ajaxSaveComponent($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-        $title = $_POST['title'] ?? '';
-        $ta_comment = $_POST['ta_comment'] ?? '';
-        $student_comment = $_POST['student_comment'] ?? '';
-        $lower_clamp = $_POST['lower_clamp'] ?? null;
-        $default = $_POST['default'] ?? null;
-        $max_value = $_POST['max_value'] ?? null;
-        $upper_clamp = $_POST['upper_clamp'] ?? null;
-        $is_itempool_linked = $_POST['is_itempool_linked'] ?? false;
-        $itempool_option = $_POST['itempool_option'] ?? null;
-
-        // Use 'page_number' since 'page' is used in the router
-        $page = $_POST['page_number'] ?? '';
-
-        // Validate required parameters
-        if ($lower_clamp === null) {
-            $this->core->getOutput()->renderJsonFail('Missing lower_clamp parameter');
-            return;
-        }
-        if ($default === null) {
-            $this->core->getOutput()->renderJsonFail('Missing default parameter');
-            return;
-        }
-        if ($max_value === null) {
-            $this->core->getOutput()->renderJsonFail('Missing max_value parameter');
-            return;
-        }
-        if ($upper_clamp === null) {
-            $this->core->getOutput()->renderJsonFail('Missing upper_clamp parameter');
-            return;
-        }
-        if ($page === '') {
-            $this->core->getOutput()->renderJsonFail('Missing page parameter');
-        }
-        if (!is_numeric($lower_clamp)) {
-            $this->core->getOutput()->renderJsonFail('Invalid lower_clamp parameter');
-            return;
-        }
-        if (!is_numeric($default)) {
-            $this->core->getOutput()->renderJsonFail('Invalid default parameter');
-            return;
-        }
-        if (!is_numeric($max_value)) {
-            $this->core->getOutput()->renderJsonFail('Invalid max_value parameter');
-            return;
-        }
-        if (!is_numeric($upper_clamp)) {
-            $this->core->getOutput()->renderJsonFail('Invalid upper_clamp parameter');
-            return;
-        }
-        if (strval(intval($page)) !== $page) {
-            $this->core->getOutput()->renderJsonFail('Invalid page parameter');
-        }
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.save_component", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to save components');
-            return;
-        }
-
-        $is_notebook_gradeable = ($gradeable->getAutogradingConfig() !== null) && $gradeable->getAutogradingConfig()->isNotebookGradeable();
-
-        if ($is_notebook_gradeable) {
-            if ($is_itempool_linked === 'true') {
-                if (!$itempool_option) {
-                    $this->core->getOutput()->renderJsonFail('Missing itempool_option parameter');
-                    return;
-                }
-            }
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $component->setTitle($title);
-            $component->setTaComment($ta_comment);
-            $component->setStudentComment($student_comment);
-            $component->setPoints([
-                'lower_clamp' => $lower_clamp,
-                'default' => $default,
-                'max_value' => $max_value,
-                'upper_clamp' => $upper_clamp
-            ]);
-            $component->setPage($page);
-            if ($is_notebook_gradeable) {
-                if ($is_itempool_linked === 'true') {
-                    $component->setIsItempoolLinked(true);
-                    $component->setItempool($itempool_option);
-                }
-                else {
-                    $component->setIsItempoolLinked(false);
-                    $component->setItempool('');
-                }
-            }
-
-            $this->core->getQueries()->saveComponent($component);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    /**
-     * Route for saving the order of components in a gradeable
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/order", methods: ["POST"])]
-    public function ajaxSaveComponentOrder($gradeable_id) {
-        // Required parameters
-        $order = json_decode($_POST['order'] ?? '[]', true);
-
-        // Validate required parameters
-        if (count($order) === 0) {
-            $this->core->getOutput()->renderJsonFail('Missing order parameter');
-            return;
-        }
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.save_component", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to save marks');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $this->saveComponentOrder($gradeable, $order);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function saveComponentOrder(Gradeable $gradeable, array $orders) {
-        foreach ($gradeable->getComponents() as $component) {
-            if (!isset($orders[$component->getId()])) {
-                throw new \InvalidArgumentException('Missing component id in order array');
-            }
-            $order = $orders[$component->getId()];
-            if (!is_int($order) || $order < 0) {
-                throw new \InvalidArgumentException('All order values must be non-negative integers');
-            }
-            $component->setOrder(intval($order));
-        }
-        $this->core->getQueries()->updateGradeable($gradeable);
-    }
-
-    /**
-     * Route for saving the page numbers of the components
-     * NOTE: the 'pages' parameter can be an associate array to set the page numbers of each component,
-     *  or a single-element array with the key 'page' of the page number to set all components' page to
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/save_pages", methods: ["POST"])]
-    public function ajaxSaveComponentPages($gradeable_id) {
-        // Required parameters
-        $pages = json_decode($_POST['pages'] ?? '[]', true);
-
-        // Validate required parameters
-        if (count($pages) === 0) {
-            $this->core->getOutput()->renderJsonFail('Missing pages parameter');
-            return;
-        }
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.save_component", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to save marks');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            if (isset($pages['page']) && count($pages) === 1) {
-                // if one page is sent, set all to that page.  This is useful
-                //  for setting the page settings to 'none' or 'student-assign'
-                $this->saveComponentsPage($gradeable, $pages['page']);
-            }
-            else {
-                $this->saveComponentPages($gradeable, $pages);
-            }
-            $this->core->getQueries()->updateGradeable($gradeable);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function saveComponentPages(Gradeable $gradeable, array $pages) {
-        foreach ($gradeable->getComponents() as $component) {
-            if (!isset($pages[$component->getId()])) {
-                throw new \InvalidArgumentException('Missing component id in pages array');
-            }
-            $page = $pages[$component->getId()];
-            if (!is_int($page)) {
-                throw new \InvalidArgumentException('All page values must be integers');
-            }
-            $component->setPage(max(intval($page), -1));
-        }
-    }
-
-    public function saveComponentsPage(Gradeable $gradeable, int $page) {
-        foreach ($gradeable->getComponents() as $component) {
-            $component->setPage(max($page, -1));
-        }
-    }
-
-    /**
-     * Route for adding a new component to a gradeable
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/new", methods:["POST"])]
-    public function ajaxAddComponent($gradeable_id) {
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        $peer = $_POST['peer'] === 'true';
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.add_component", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to add components');
-            return;
-        }
-
-        try {
-            $page = $gradeable->isPdfUpload() ? ($gradeable->isStudentPdfUpload() ? Component::PDF_PAGE_STUDENT : 1) : Component::PDF_PAGE_NONE;
-
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $component = $gradeable->addComponent(
-                'Problem ' . strval(count($gradeable->getComponents()) + 1),
-                '',
-                '',
-                0,
-                0,
-                0,
-                0,
-                false,
-                $peer,
-                $page
-            );
-            $component->addMark('No Credit', 0.0, false);
-            $this->core->getQueries()->updateGradeable($gradeable);
-            $this->core->getOutput()->renderJsonSuccess(['component_id' => $component->getId()]);
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    /**
-     * Route for deleting a component from a gradeable
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/delete", methods: ["POST"])]
-    public function ajaxDeleteComponent($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.delete_component", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to delete components');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $gradeable->deleteComponent($component);
-            $this->core->getQueries()->updateGradeable($gradeable);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    /**
-     * Route for saving a mark's title/point value
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/marks/save", methods: ["POST"])]
-    public function ajaxSaveMark($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-        $mark_id = $_POST['mark_id'] ?? '';
-        $points = $_POST['points'] ?? '';
-        $title = $_POST['title'] ?? null;
-        $publish = ($_POST['publish'] ?? 'false') === 'true';
-
-        // Validate required parameters
-        if ($title === null) {
-            $this->core->getOutput()->renderJsonFail('Missing title parameter');
-            return;
-        }
-        if ($points === null) {
-            $this->core->getOutput()->renderJsonFail('Missing points parameter');
-            return;
-        }
-        if (!is_numeric($points)) {
-            $this->core->getOutput()->renderJsonFail('Invalid points parameter');
-            return;
-        }
-
-        $points = floatval($points);
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // get the mark
-        $mark = $this->tryGetMark($component, $mark_id);
-        if ($mark === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.save_mark", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to save marks');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $this->saveMark($mark, $points, $title, $publish);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function saveMark(Mark $mark, float $points, string $title, bool $publish) {
-        if ($mark->getPoints() !== $points) {
-            $mark->setPoints($points);
-        }
-        if ($mark->getTitle() !== $title) {
-            $mark->setTitle($title);
-        }
-        if ($mark->isPublish() !== $publish) {
-            $mark->setPublish($publish);
-        }
-        $this->core->getQueries()->updateGradeable($mark->getComponent()->getGradeable());
-    }
-
-    /**
-     * Route for saving a the order of marks in a component
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/marks/save_order", methods: ["POST"])]
-    public function ajaxSaveMarkOrder($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-        $order = json_decode($_POST['order'] ?? '[]', true);
-
-        // Validate required parameters
-        if (count($order) === 0) {
-            $this->core->getOutput()->renderJsonFail('Missing order parameter');
-            return;
-        }
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.save_mark", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to save marks');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $this->saveMarkOrder($component, $order);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function saveMarkOrder(Component $component, array $orders) {
-        foreach ($component->getMarks() as $mark) {
-            if (!isset($orders[$mark->getId()])) {
-                throw new \InvalidArgumentException('Missing mark id in order array');
-            }
-            $order = $orders[$mark->getId()];
-            if (!is_int($order) || $order < 0) {
-                throw new \InvalidArgumentException('All order values must be non-negative integers');
-            }
-            $mark->setOrder(intval($order));
-        }
-        $this->core->getQueries()->saveComponent($component);
-    }
-
     /**
      * Route for getting the student's program output for the diff-viewer
      */
@@ -3262,132 +1807,6 @@ class ElectronicGraderController extends AbstractController {
             $this->core->getOutput()->renderJsonError($e->getMessage());
         }
     }
-
-    /**
-     * Route for adding a mark to a component
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/marks/add", methods: ["POST"])]
-    public function ajaxAddNewMark($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-        $points = $_POST['points'] ?? '';
-        $title = $_POST['title'] ?? null;
-        $publish = ($_POST['publish'] ?? 'false') === 'true';
-
-        // Validate required parameters
-        if ($title === null) {
-            $this->core->getOutput()->renderJsonFail('Missing title parameter');
-            return;
-        }
-        if ($points === null) {
-            $this->core->getOutput()->renderJsonFail('Missing points parameter');
-            return;
-        }
-        if (!is_numeric($points)) {
-            $this->core->getOutput()->renderJsonFail('Invalid points parameter');
-            return;
-        }
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.add_new_mark", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to add mark');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $mark = $this->addNewMark($component, $title, $points, $publish);
-            $this->core->getOutput()->renderJsonSuccess(['mark_id' => $mark->getId()]);
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function addNewMark(Component $component, string $title, float $points, bool $publish) {
-        $mark = $component->addMark($title, $points, $publish);
-        $this->core->getQueries()->saveComponent($component);
-        return $mark;
-    }
-
-    /**
-     * Route for deleting a mark from a component
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/marks/delete", methods: ["POST"])]
-    public function ajaxDeleteMark($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-        $mark_id = $_POST['mark_id'] ?? '';
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // get the mark
-        $mark = $this->tryGetMark($component, $mark_id);
-        if ($mark === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.delete_mark", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to delete marks');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $this->deleteMark($mark);
-            $this->core->getOutput()->renderJsonSuccess();
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    public function deleteMark(Mark $mark) {
-        $mark->getComponent()->deleteMark($mark);
-        $this->core->getQueries()->saveComponent($mark->getComponent());
-    }
-
     /**
      * Route for saving the general comment for the gradeable
      */
@@ -3579,138 +1998,6 @@ class ElectronicGraderController extends AbstractController {
         // Once we've parsed the inputs and checked permissions, perform the operation
         $this->core->getOutput()->renderJsonSuccess($ta_graded_gradeable->getOverallComments($commenter));
     }
-
-    /**
-     * Route for getting all submitters that received a mark and stats about that mark
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/components/marks/stats", methods: ["POST"])]
-    public function ajaxGetMarkStats($gradeable_id) {
-        // Required parameters
-        $component_id = $_POST['component_id'] ?? '';
-        $mark_id = $_POST['mark_id'] ?? '';
-
-        $grader = $this->core->getUser();
-
-        // Get the gradeable
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return;
-        }
-
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-
-        // get the component
-        $component = $this->tryGetComponent($gradeable, $component_id);
-        if ($component === false) {
-            return;
-        }
-
-        // get the mark
-        $mark = $this->tryGetMark($component, $mark_id);
-        if ($mark === false) {
-            return;
-        }
-
-        // checks if user has permission
-        if (!$this->core->getAccess()->canI("grading.electronic.get_marked_users", ["gradeable" => $gradeable])) {
-            $this->core->getOutput()->renderJsonFail('Insufficient permissions to view marked users');
-            return;
-        }
-
-        try {
-            // Once we've parsed the inputs and checked permissions, perform the operation
-            $results = $this->getMarkStats($mark, $grader);
-            $this->core->getOutput()->renderJsonSuccess($results);
-        }
-        catch (\InvalidArgumentException $e) {
-            $this->core->getOutput()->renderJsonFail($e->getMessage());
-        }
-        catch (\Exception $e) {
-            $this->core->getOutput()->renderJsonError($e->getMessage());
-        }
-    }
-
-    private function getMarkStats(Mark $mark, User $grader) {
-        $gradeable = $mark->getComponent()->getGradeable();
-
-        $anon = $this->amIBlindGrading($gradeable, $grader, false);
-
-        $section_submitter_ids = $this->core->getQueries()->getSubmittersWhoGotMarkBySection($mark, $grader, $gradeable, $anon);
-        $all_submitter_ids     = $this->core->getQueries()->getAllSubmittersWhoGotMark($mark, $anon);
-
-        // Show all submitters if grader has permissions, otherwise just section submitters
-        $submitter_ids = ($grader->accessFullGrading() ? $all_submitter_ids : $section_submitter_ids);
-
-        if ($gradeable->isTeamAssignment()) {
-            $submitter_anon_ids = ($anon != 'unblind') ? $submitter_ids : $this->core->getQueries()->getTeamAnonId($submitter_ids);
-        }
-        else {
-            $submitter_anon_ids = ($anon != 'unblind') ? $submitter_ids : $this->core->getQueries()->getAnonId($submitter_ids, $gradeable->getId());
-        }
-
-        $section_graded_component_count = 0;
-        $section_total_component_count  = 0;
-        $total_graded_component_count   = 0;
-        $total_total_component_count    = 0;
-
-        $this->getStats($gradeable, $grader, true, $total_graded_component_count, $total_total_component_count);
-        $this->getStats($gradeable, $grader, false, $section_graded_component_count, $section_total_component_count);
-
-        return [
-            'section_submitter_count' => count($section_submitter_ids),
-            'total_submitter_count'   => count($all_submitter_ids),
-            'section_graded_component_count' => $section_graded_component_count,
-            'total_graded_component_count'   => $total_graded_component_count,
-            'section_total_component_count' => $section_total_component_count,
-            'total_total_component_count'   => $total_total_component_count,
-            'submitter_ids' => $submitter_ids,
-            'submitter_anon_ids' => $submitter_anon_ids
-        ];
-    }
-
-    /**
-     * Gets... stats
-     * @param Gradeable $gradeable
-     * @param User      $grader
-     * @param bool      $full_stats
-     */
-    private function getStats(Gradeable $gradeable, User $grader, bool $full_stats, &$total_graded, &$total_total) {
-        $num_components = $this->core->getQueries()->getTotalComponentCount($gradeable->getId());
-        $sections = [];
-        if ($full_stats) {
-            $sections = $this->core->getQueries()->getAllSectionsForGradeable($gradeable);
-        }
-        elseif ($gradeable->isGradeByRegistration()) {
-            $sections = $grader->getGradingRegistrationSections();
-        }
-        else {
-            $sections = $this->core->getQueries()->getRotatingSectionsForGradeableAndUser($gradeable->getId(), $grader->getId());
-        }
-
-        $include_withdrawn_students = ($_COOKIE['include_withdrawn_students'] ?? 'omit') === 'include';
-        $section_key = ($gradeable->isGradeByRegistration() ? 'registration_section' : 'rotating_section');
-        $total_users       = [];
-        $graded_components = [];
-        $ta_graded_components = [];
-        if (count($sections) > 0) {
-            $total_users = ($gradeable->isTeamAssignment()) ?
-            $this->core->getQueries()->getTotalTeamCountByGradingSections($gradeable->getId(), $sections, $section_key) :
-            $this->core->getQueries()->getTotalUserCountByGradingSections($sections, $section_key, $include_withdrawn_students);
-            $graded_components = $this->core->getQueries()->getGradedComponentsCountByGradingSections($gradeable->getId(), $sections, $section_key, $gradeable->isTeamAssignment(), $include_withdrawn_students);
-        }
-
-        foreach ($graded_components as $key => $value) {
-            $total_graded += intval($value);
-        }
-        foreach ($total_users as $key => $value) {
-            $total_total += $value * $num_components;
-        }
-    }
-
     /**
      * @param Gradeable $gradeable
      * @param string $who_id
@@ -3820,74 +2107,5 @@ class ElectronicGraderController extends AbstractController {
             "component_id" => $component_id,
             "itempool_item" => $solution_row['itempool_item'],
         ]) : JsonResponse::getErrorResponse($error);
-    }
-
-    /**
-     */
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/feedback/set", methods: ["POST"])]
-    public function ajaxSetPeerFeedback($gradeable_id) {
-        $grader_id = $_POST['grader_id'] ?? '';
-        $user_id = $_POST['user_id'] ?? '';
-        $feedback = $_POST['feedback'];
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            return null;
-        }
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonError("This gradeable is not an electronic file gradeable");
-            return null;
-        }
-        $graded_gradeable = $this->tryGetGradedGradeable($gradeable, $user_id)->getGradeableId() == $gradeable_id;
-        if ($graded_gradeable === false) {
-            return null;
-        }
-        if ($gradeable->setPeerFeedback($this->core->getQueries()->getUserFromAnon($grader_id, $gradeable_id)[$grader_id], $user_id, $feedback)) {
-            $this->core->getOutput()->renderJsonSuccess("Feedback successfully uploaded");
-        }
-        else {
-            $this->core->getOutput()->renderJsonError("Failed to save feedback");
-        }
-        return true;
-    }
-
-    #[AccessControl(role: "FULL_ACCESS_GRADER")]
-    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/grading/clear_peer_marks", methods: ["POST"])]
-    public function ajaxClearPeerMarks($gradeable_id) {
-        $submitter_id = $_POST['submitter_id'] ?? '';
-        $peer_id = $_POST['peer_id'] ?? '';
-        $gradeable = $this->tryGetGradeable($gradeable_id);
-        if ($gradeable === false) {
-            $this->core->getOutput()->renderJsonFail('Could not fetch gradeable');
-            return;
-        }
-        // Make sure this gradeable is an electronic file gradeable
-        if ($gradeable->getType() !== GradeableType::ELECTRONIC_FILE) {
-            $this->core->getOutput()->renderJsonFail('This gradeable is not an electronic file gradeable');
-            return;
-        }
-        $graded_gradeable = $this->tryGetGradedGradeable($gradeable, $submitter_id);
-        if ($graded_gradeable === false) {
-            $this->core->getOutput()->renderJsonFail('Could not fetch graded gradeable');
-            return;
-        }
-        $ta_graded_gradeable = $graded_gradeable->getOrCreateTaGradedGradeable();
-        foreach ($ta_graded_gradeable->getGradedComponentContainers() as $container) {
-            $component = $container->getComponent();
-            $ta_graded_gradeable->deleteGradedComponent($component, $this->core->getQueries()->getUserById($peer_id));
-        }
-        $ta_graded_gradeable->removeOverallComment($peer_id);
-        $this->core->getQueries()->deleteOverallComment($gradeable_id, $peer_id, $graded_gradeable->getSubmitter()->getId());
-        $this->core->getQueries()->deleteTaGradedGradeableByIds($gradeable_id, $peer_id);
-        $ta_graded_gradeable->resetUserViewedDate();
-
-        // Finally, save the graded gradeable
-        $this->core->getQueries()->saveTaGradedGradeable($ta_graded_gradeable);
-        $submitter = $ta_graded_gradeable->getGradedGradeable()->getSubmitter();
-        if ($submitter->isTeam()) {
-            $this->core->getQueries()->clearTeamViewedTime($submitter->getId());
-        }
-        $this->core->getOutput()->renderJsonSuccess("Marks removed successfully!");
-        return true;
     }
 }
